# docs/api/openapi.yaml
components:
  schemas:
...
    CompleteUploadRequest:
      properties:
        checksumSha256:
          description: Optional. When supplied it is verified against what object storage actually holds.
          pattern: "^[A-Fa-f0-9]{64}$"
          type: string
      type: object
...
    DocumentResponse:
      properties:
        applicationId:
          format: uuid
          type: string
        createdAt:
          format: date-time
          type: string
        documentId:
          format: uuid
          type: string
        documentType:
          description: "PROOF_OF_IDENTITY and PROOF_OF_INCOME are mandatory: an application cannot be submitted until both have reached CLEAN."
          enum:
          - PROOF_OF_IDENTITY
          - PROOF_OF_INCOME
          - PROOF_OF_ADDRESS
          - BANK_STATEMENT
          type: string
        reasonCode:
          type: string
        sizeBytes:
          format: int64
          type: integer
        status:
          description: Only CLEAN satisfies a mandatory requirement. Until then the object sits in a quarantine prefix and is never served.
          enum:
          - PENDING_UPLOAD
          - UPLOADED
          - SCANNING
          - CLEAN
          - REJECTED
          type: string
        updatedAt:
          format: date-time
          type: string
      type: object
...
    UploadRequest:
      properties:
        contentType:
          description: The presigned URL is BOUND to this value. An upload whose Content-Type differs is refused by object storage.
          example: application/pdf
          minLength: 1
          type: string
        documentType:
          minLength: 1
          pattern: ^(PROOF_OF_IDENTITY|PROOF_OF_INCOME|PROOF_OF_ADDRESS|BANK_STATEMENT)$
          type: string
        sizeBytes:
          description: Declared size. Verified against what object storage actually holds.
          format: int64
          minimum: 0
          type: integer
      required:
      - contentType
      - documentType
      type: object
    UploadTicketResponse:
      properties:
        documentId:
          format: uuid
          type: string
        expiresAt:
          description: Deliberately short-lived.
          format: date-time
          type: string
        requiredHeaders:
          additionalProperties:
            type: string
          description: "Headers that must be sent verbatim on the PUT. The signature covers them, so the upload is refused if any differs."
          type: object
        uploadUrl:
          description: |
            **A bearer credential.** Anyone holding this URL can write the object until it expires.

            Never log it, never cache it, never put it in a URL shortener or an analytics payload. The response is returned with `Cache-Control: no-store` for the same reason.
          format: uri
          type: string
      type: object
  securitySchemes:
    oauth2:
      description: |
        An OAuth2 bearer token, validated at API Gateway **and again by each service**. The second validation is what protects the services if anything is ever deployed with a route that bypasses the gateway.

        Partners additionally present an API key. The key is what a usage plan throttles on — a capacity identity, never an authentication one — so a key alone grants nothing.
      flows:
        clientCredentials:
          scopes:
            applications:read: Read applications and their status
            applications:write: "Create, update and submit applications"
            documents:read: List documents on an application
            documents:write: Request upload slots and confirm uploads
            manual-review:decide: Record a reviewer's decision
            manual-review:read: Read the manual review queue
          tokenUrl: https://auth.example.com/oauth2/token
      type: oauth2
info:
  contact:
    name: Repository
    url: https://example.com/aws-loan-origination-platform
  description: |
    ## What this is

    A **reference implementation**. It has never been deployed, it has never processed real data, and every external check behind it — KYC, AML, fraud, credit scoring, malware scanning — is a **simulation** that contacts no real provider.

    ## Two services, one API surface

    These paths are served by two independently deployed services behind a single API Gateway. `/v1/applications/{applicationId}/documents/**` belongs to the document service; everything else belongs to the application service. Callers should not depend on that split — it is recorded because it explains why a document can be accepted a moment before submission will accept it.

    ## Idempotency

    `POST /v1/applications` and `POST .../submit` accept an `Idempotency-Key` header, and submission is the one that most needs it: a client that times out and retries must not submit twice.

    | Situation | Result |
    |---|---|
    | Same key, same request | The original response, with `Idempotent-Replay: true` |
    | Same key, different request | `409` with `IDEMPOTENCY_KEY_REUSED` |
    | Same key, original still in flight | `409` with `IDEMPOTENCY_KEY_IN_USE` |

...

    ## Errors

...

    Branch on `errorCode`, never on `detail`, which is prose and may be reworded.

    ## Personal data

...

...
  license:
    identifier: Apache-2.0
    name: Apache-2.0
...
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/applications/{applicationId}/documents:
    get:
      description: Each carries its scan status. Only `CLEAN` satisfies a mandatory requirement.
      operationId: listDocuments
      parameters:
      - in: path
        name: applicationId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: "#/components/schemas/DocumentResponse"
                type: array
          description: "The documents, with their scan status."
        "401":
          description: "No token, or a token that is not valid here."
        "403":
          description: The token lacks `documents:read`.
      summary: The documents attached to an application
      tags:
      - Documents
  /v1/applications/{applicationId}/documents/upload-requests:
    post:
      description: |
        Returns a short-lived presigned `PUT` URL. The client uploads **directly to object storage** — the bytes never pass through this API.

        There is deliberately no filename field. A client-supplied filename is user-controlled text that has to go somewhere, and every destination is wrong: in the object key it leaks through access logs and inventory reports, in the database it is personal data with no purpose, and in a response header it is a content-disposition injection.

        **The response carries a bearer credential.** It is returned with `Cache-Control: no-store` — not merely `no-cache` — because a presigned URL sitting in a shared cache, a proxy, or a browser's disk cache is a write credential handed to whoever reads that cache next. Do not log it.
      operationId: requestUpload
      parameters:
      - in: path
        name: applicationId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UploadTicketResponse"
          description: "An upload slot. Always `Cache-Control: no-store`."
        "400":
          description: The request failed validation.
        "403":
          description: The token lacks `documents:write`.
      summary: Request an upload slot
      tags:
      - Documents
  /v1/applications/{applicationId}/documents/{documentId}/complete:
    post:
      description: |
        The service does not take the client's word for it. It asks object storage what is actually at the key and verifies size and content type against what was declared, then queues the object for scanning.

        The document becomes usable only once it reaches `CLEAN`. Until then it sits in a quarantine prefix and cannot satisfy a mandatory requirement.
      operationId: completeUpload
      parameters:
      - in: path
        name: applicationId
        required: true
        schema:
          type: string
      - in: path
        name: documentId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteUploadRequest"
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DocumentResponse"
          description: The document's state after verification.
        "404":
          description: No such document.
        "422":
          description: "What is at the key does not match what was declared, or the document is not awaiting an upload."
      summary: Confirm that an upload finished
      tags:
      - Documents
...
security:
- oauth2: []
servers:
...